Makes SecretBundleConfigSource public
